+++ image_builder/image_builder/docker/template.py
@@ -1,5 +1,4 @@
 from pathlib import Path
-from typing import Optional, Dict
 
 from jinja2 import Template
 
@@ -11,22 +10,22 @@ def template_dockerfile(
     manifest: Manifest,
     template_path: Path,
     dockerfile_path: Path,
-    base_image: Optional[str],
     git_version: str,
-    deployed_by_racetrack_version: Optional[str],
-    job_type_version: Optional[str],
-    env_vars: Dict[str, str],
+    racetrack_version: str | None,
+    job_type_version: str | None,
+    env_vars: dict[str, str],
+    base_image: str | None = None,
 ):
     """
     Create Dockerfile from Jinja template and manifest data
     :param manifest: Job manifest data
     :param template_path: Path to Jinja template file
     :param dockerfile_path: Path to output Dockerfile that will be created from template
-    :param base_image: Full name of base image
     :param git_version: version name from Job git history
-    :param deployed_by_racetrack_version: Version of Racetrack the Job was deployed with
+    :param racetrack_version: Version of Racetrack the Job was deployed with
     :param job_type_version: Version of Job Type used to build the Job's image
     :param env_vars: environment variables that should be set during building
+    :param base_image: (Deprecated) name of the base docker image
     """
     template_content = Path(template_path).read_text()
     template = Template(template_content)
@@ -36,9 +35,10 @@ def template_dockerfile(
         'base_image': base_image,
         'resource_name': resource_name,
         'git_version': git_version,
-        'deployed_by_racetrack_version': deployed_by_racetrack_version,
+        'deployed_by_racetrack_version': racetrack_version,
         'job_type_version': job_type_version,
         'env_vars': env_vars,
+        'manifest_jobtype_extra': manifest.get_jobtype_extra() or {},
     }
     templated = template.render(**render_vars)
     Path(dockerfile_path).write_text(templated)