+++ racetrack_commons/racetrack_commons/deploy/job_type.py
@@ -1,30 +1,35 @@
 from __future__ import annotations
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Union
 
 import backoff
 
 from racetrack_client.log.context_error import wrap_context
+from racetrack_client.log.logs import get_logger
 from racetrack_client.plugin.plugin_manifest import PluginManifest
 from racetrack_client.utils.semver import SemanticVersion, SemanticVersionPattern
 from racetrack_commons.plugin.core import PluginCore
 from racetrack_commons.plugin.engine import PluginEngine
+from racetrack_commons.plugin.plugin_data import PluginData
 
-
-JobTypeImagePaths = Union[list[tuple[Path, Path]], tuple[Path, Path]]
+logger = get_logger(__name__)
 
 
 @dataclass
 class JobType:
     lang_name: str
     version: str  # semantic version of the job type
-    base_image_paths: list[Path]  # paths to base Dockerfiles (for each container)
-    template_paths: list[Path]  # paths to job template Dockerfiles (for each container)
+    template_paths: list[str | None]  # relative path names to job template Dockerfiles (for each container)
+    jobtype_dir: Path
+    base_image_paths: list[Path | None]  # Deprecated
 
     @property
-    def full_name(self):
+    def full_name(self) -> str:
         return f'{self.lang_name}:{self.version}'
+
+    @property
+    def containers_num(self) -> int:
+        return len(self.template_paths)
 
 
 @backoff.on_exception(backoff.fibo, AssertionError, max_value=1, max_time=5, jitter=None, logger=None)
@@ -82,58 +87,73 @@ def gather_job_types(
 ) -> dict[str, JobType]:
     """
     Load job types from plugins.
-    Return job name (with version) -> (base image name, dockerfile template path)
+    Return dictionary: job name (with version) mapped to JobType data
     """
     job_types: dict[str, JobType] = {}
 
-    plugin_results: list[dict[str, JobTypeImagePaths]] = plugin_engine.invoke_plugin_hook(PluginCore.job_types)
-    for plugin_job_types in plugin_results:
-        if plugin_job_types:
-            for job_full_name, job_data in plugin_job_types.items():
+    plugin_data: PluginData
+    plugin_job_types: dict[str, list[str | Path]]
+    for plugin_data, plugin_job_types in plugin_engine.invoke_hook_with_origin(PluginCore.job_types):
+        if not plugin_job_types:
+            continue
 
-                if isinstance(job_data, tuple):
-                    base_image_paths = [job_data[0]]
-                    template_paths = [job_data[1]]
-                elif isinstance(job_data, list):
-                    base_image_paths = [item[0] for item in job_data]
-                    template_paths = [item[1] for item in job_data]
+        for job_type_key, job_type_value in plugin_job_types.items():
+
+            template_paths: list[str | None]
+            base_image_paths: list[Path | None] = []  # Deprecated, kept for backwards compatibility
+            if isinstance(job_type_value, list):
+                if all(isinstance(item, tuple) for item in job_type_value):
+                    logger.warning('Using deprecated base images. Please use a single job template instead.')
+                    base_image_paths = [Path(item[0]) if item[0] else None
+                                        for item in job_type_value]
+                    template_paths = [Path(item[1]).as_posix() if item[1] else None
+                                      for item in job_type_value]
                 else:
-                    raise RuntimeError(f'Invalid job type data. It should be list[tuple[Path, Path]], was {type(job_data)}')
+                    for item in job_type_value:
+                        if item is not None:
+                            assert isinstance(item, str), f'Invalid dockerfile template path of a job type. Expected str, but found {type(item)}'
+                    template_paths = job_type_value
+            elif isinstance(job_type_value, str):
+                template_paths = [job_type_value]
+            elif isinstance(job_type_value, tuple):
+                logger.warning('Using deprecated base images. Please use a single job template instead.')
+                base_image_paths = [Path(job_type_value[0])]
+                template_paths = [Path(job_type_value[1]).as_posix()]
+            else:
+                raise RuntimeError(f'Invalid job type data. Expected list[str], was {type(job_type_value)}')
+            assert len(template_paths) > 0, 'list of job type template paths can not be empty'
 
-                name_parts = job_full_name.split(':')
-                assert len(name_parts) == 2, f'job type {job_full_name} should have the version defined (name:version)'
-                lang_name = name_parts[0]
-                lang_version = name_parts[1]
-                job_type_version = JobType(
-                    lang_name=lang_name,
-                    version=lang_version,
-                    base_image_paths=base_image_paths,
-                    template_paths=template_paths,
-                )
-                job_types[job_full_name] = job_type_version
+            name_parts = job_type_key.split(':')
+            assert len(name_parts) == 2, f'job type {job_type_key} should have the version defined (name:version)'
+            lang_name = name_parts[0]
+            lang_version = name_parts[1]
+            job_type_version = JobType(
+                lang_name=lang_name,
+                version=lang_version,
+                template_paths=template_paths,
+                jobtype_dir=plugin_data.plugin_dir,
+                base_image_paths=base_image_paths,
+            )
+            job_types[job_type_key] = job_type_version
 
     return job_types
 
 
 def _validate_job_type(job_type: JobType):
-    _validate_dockerfile_paths(job_type.base_image_paths, job_type.template_paths, job_type.full_name)
+    template_paths: list[str | None] = job_type.template_paths
+    job_full_name: str = job_type.full_name
+    assert len(template_paths) > 0, 'Job type should have non-empty list of template images'
 
-
-def _validate_dockerfile_paths(base_image_paths: list[Path], template_paths: list[Path], job_full_name: str):
-    assert len(base_image_paths) > 0, 'Job type should have non-empty list of base images'
-
-    for i, base_image_path in enumerate(base_image_paths):
-        template_path = template_paths[i]
-        if base_image_path is not None:
-            assert base_image_path.is_file(), f'cannot find base image Dockerfile for {job_full_name} language wrapper: {base_image_path}'
+    for i, template_path in enumerate(template_paths):
         if template_path is not None:
-            assert template_path.is_file(), f'cannot find Dockerfile template for {job_full_name} language wrapper: {template_path}'
+            template_full_path = job_type.jobtype_dir / template_path
+            assert template_full_path.is_file(), f'cannot find Dockerfile template for {job_full_name} job type: {template_full_path}'
 
 
 def list_jobtype_names_of_plugins(plugin_engine: PluginEngine) -> list[tuple[PluginManifest, str]]:
     entries: list[tuple[PluginManifest, str]] = []
-    for plugin_manifest, result in plugin_engine.invoke_associated_plugin_hook(PluginCore.job_types):
+    for plugin_data, result in plugin_engine.invoke_hook_with_origin(PluginCore.job_types):
         if result:
             for jobtype_name in result.keys():
-                entries.append((plugin_manifest, jobtype_name))
+                entries.append((plugin_data.plugin_manifest, jobtype_name))
     return entries